Database

+++ admin/processing/fetch_ban.php
@@ -6,16 +6,21 @@
     if(isset($_POST['check'])){
         $query = "SELECT role FROM users WHERE id = {$user_id}";
         if($result = mysqli_query($link,$query)){
-            $row = mysqli_fetch_assoc($result);
-            $current_role = $row['role'];
-            $logged_role = $_SESSION['role'];
-            if($current_role != $logged_role){
-                if($current_role > 1){
-                    $query = "DELETE FROM notifications WHERE type = 1 AND receipient = {$user_id}";
-                    mysqli_query($link,$query);
-                    $query = "INSERT INTO notifications VALUES ({$user_id},4,NULL,0)";
-                    mysqli_query($link,$query);
+            if(mysqli_num_rows($result) > 0){
+                $row = mysqli_fetch_assoc($result);
+                $current_role = $row['role'];
+                $logged_role = $_SESSION['role'];
+                if($current_role != $logged_role){
+                    if($current_role > 1){
+                        $query = "DELETE FROM notifications WHERE type = 1 AND receipient = {$user_id}";
+                        mysqli_query($link,$query);
+                        $query = "INSERT INTO notifications VALUES ({$user_id},4,NULL,0)";
+                        mysqli_query($link,$query);
+                    }
+                    header('location:../logout.php');
+                    die;
                 }
+            }else{
                 header('location:../logout.php');
                 die;
             }
